fix(slack): move SlackActionClientOpts to actions-client to clear madge cycle
Define SlackActionClientOpts in actions-client.ts (the leaf module) so
actions-bookmarks.ts imports both getClient and the opts type from one
leaf, dropping the type-only edge back to actions.ts that madge flagged
as a cycle. actions.ts re-imports getClient/resolveToken/opts locally
and re-exports them through its barrel.
