+++ extensions/slack/src/actions-bookmarks.ts
@@ -1,8 +1,7 @@
 // Slack channel bookmark operations, split from actions.ts to keep that file
 // under the repo max-lines limit. Re-exported by actions.ts so the lazy
 // action-runtime module loader resolves them through the actions.js barrel.
-import { getClient } from "./actions-client.js";
-import type { SlackActionClientOpts } from "./actions.js";
+import { getClient, type SlackActionClientOpts } from "./actions-client.js";
 
 export type SlackBookmark = {
   id?: string;
